refactor(uBikeTable): 表格拆成子元件
