+++ src/views/uBikeTable/uBikeTable.vue
@@ -2,8 +2,9 @@
 import { ref, computed, watch } from 'vue';
 import "bootstrap/dist/css/bootstrap.css";
 import '@fortawesome/fontawesome-free/css/all.css';
-import pagination from './components/pagination.vue'
 import search from './components/search.vue'
+import uBikeTable from './components/uBikeTable.vue'
+import pagination from './components/pagination.vue'
 
 // 修改這份 YouBike 即時資訊表：
 // 1. 將搜尋的部分拆出來變成子元件 `uBikeTable/components/search.vue`
@@ -19,11 +20,6 @@ import search from './components/search.vue'
 // lat：緯度、 lng：經度、 ar：地(中文)、 sareaen：場站區域(英文)、
 // snaen：場站名稱(英文)、 aren：地址(英文)、 bemp：空位數量、 act：全站禁用狀態
 
-// 目前的排序選項
-const currentSort = ref('sno');
-// 是否為降冪排序
-const isSortDesc = ref(false);
-
 // 所有站點資料
 const uBikeStops = ref([]);
 // 搜尋文字
@@ -34,7 +30,6 @@ const currentPage = ref(1);
 const COUNT_OF_PAGE = 20;
 // 頁碼最多顯示幾頁
 const PAGINATION_MAX = 10;
-
 
 fetch('https://tcgbusfs.blob.core.windows.net/dotapp/youbike/v2/youbike_immediate.json')
   .then(res => res.text())
@@ -58,24 +53,14 @@ const filtedUbikeStops = computed(() => {
     : uBikeStops.value.filter(d => d.sna.includes(searchText.value));
 });
 
-// 排序後的站點資料
-const sortedUbikeStops = computed(() => {
-  const filtedStops = [...filtedUbikeStops.value];
-
-  return isSortDesc.value
-    ? filtedStops.sort((a, b) => b[currentSort.value] - a[currentSort.value])
-    : filtedStops.sort((a, b) => a[currentSort.value] - b[currentSort.value]);
-});
-
-
 // 分頁後的站點資料
 const slicedUbikeStops = computed(() => {
   const start = (currentPage.value - 1) * COUNT_OF_PAGE;
   const end =
-    start + COUNT_OF_PAGE <= sortedUbikeStops.value.length
+    start + COUNT_OF_PAGE <= filtedUbikeStops.value.length
       ? start + COUNT_OF_PAGE
-      : sortedUbikeStops.value.length;
-  return sortedUbikeStops.value.slice(start, end);
+      : filtedUbikeStops.value.length;
+  return filtedUbikeStops.value.slice(start, end);
 });
 
 // 總頁數
@@ -111,71 +96,15 @@ const setPage = page => {
   currentPage.value = page;
 };
 
-// 指定排序
-const setSort = sortType => {
-  if (sortType === currentSort.value) {
-    isSortDesc.value = !isSortDesc.value;
-  } else {
-    currentSort.value = sortType;
-    isSortDesc.value = false;
-  }
-};
-
-// 關鍵字 Highlight
-const keywordsHighlight = (text, keyword) => {
-  if(keyword === '') return text;
-  const reg = new RegExp(keyword, 'gi');
-  return text.replace(reg, `<span style="color: red;">${keyword}</span>`);
-};
 </script>
 
 <template>
   <div class="app">
     <search :internalSearchText="searchText" @setSearchState="handleSearchState"></search>
-    <table class="table table-striped">
-      <thead>
-        <tr>
-          <th @click="setSort('sno')">
-            #
-            <span v-show="currentSort === 'sno'">
-              <i class="fa" :class="isSortDesc ? 'fa-sort-desc' : 'fa-sort-asc'" aria-hidden="true"></i>
-            </span>
-          </th>
-          <th>
-            場站名稱
-          </th>
-          <th>
-            場站區域
-          </th>
-          <th @click="setSort('sbi')" class="pointer">
-            目前可用車輛
-            <span v-show="currentSort === 'sbi'">
-              <i class="fa" :class="isSortDesc ? 'fa-sort-desc' : 'fa-sort-asc'" aria-hidden="true"></i>
-            </span>
-          </th>
-          <th @click="setSort('tot')" class="pointer">
-            總停車格
-            <span v-show="currentSort === 'tot'">
-              <i class="fa" :class="isSortDesc ? 'fa-sort-desc' : 'fa-sort-asc'" aria-hidden="true"></i>
-            </span>
-          </th>
-          <th>
-            資料更新時間
-          </th>
-        </tr>
-      </thead>
-      <tbody>
-        <!-- 替換成 slicedUbikeStops -->
-        <tr v-for="s in slicedUbikeStops" :key="s.sno">
-          <td>{{ s.sno }}</td>
-          <td v-html="keywordsHighlight(s.sna, searchText)"></td>
-          <td>{{ s.sarea }}</td>
-          <td>{{ s.sbi }}</td>
-          <td>{{ s.tot }}</td>
-          <td>{{ (s.mday) }}</td>
-        </tr>
-      </tbody>
-    </table>
+    <uBikeTable
+      :internalSearchText="searchText"
+      :slicedUbikeStops="slicedUbikeStops"
+    />
   </div>
 
   <!-- 頁籤 -->
